fix(migrations): Check if exists table users

+++ database/migrations/2023_07_20_121725_add_role_id_and_avatar_to_users_table.php
@@ -13,12 +13,14 @@ return new class extends Migration
      */
     public function up()
     {
-        Schema::table('users', function (Blueprint $table) {
-            $table->string("avatar")->nullable();
+        if (Schema::hasTable('users')) {
+            Schema::table('users', function (Blueprint $table) {
+                $table->string("avatar")->nullable();
 
-            $table->unsignedBigInteger('role_id')->nullable();
-            $table->foreign('role_id')->references('id')->on('roles')->onDelete('cascade');
-        });
+                $table->unsignedBigInteger('role_id')->nullable();
+                $table->foreign('role_id')->references('id')->on('roles')->onDelete('cascade');
+            });
+        }
     }
 
     /**
@@ -28,11 +30,16 @@ return new class extends Migration
      */
     public function down()
     {
-        Schema::table('users', function (Blueprint $table) {
-            $table->dropColumn("avatar");
+        if (Schema::hasTable('users')) {
 
-            $table->dropForeign(['role_id']);
-            $table->dropColumn('role_id');
-        });
+            Schema::table('users', function (Blueprint $table) {
+                $table->dropColumn("avatar");
+
+                $table->dropForeign(['role_id']);
+                $table->dropColumn('role_id');
+            });
+
+        }
+
     }
 };
